started refactoring pcm <-> javaeditor tests to view-based
