+++ tests/tools.vitruv.applications.pcmjava.javaeditor.tests/META-INF/MANIFEST.MF
@@ -17,8 +17,9 @@ Require-Bundle: org.apache.log4j,
  edu.kit.ipd.sdq.commons.util.emf,
  edu.kit.ipd.sdq.commons.util.java,
  tools.vitruv.testutils.vsum,
+ tools.vitruv.framework.views,
  tools.vitruv.applications.util.temporary,
  tools.vitruv.applications.pcmjava,
  tools.vitruv.applications.pcmjava.tests
-Export-Package: tools.vitruv.applications.pcmjava.javaeditor.java2pcm
+Export-Package: tools.vitruv.applications.pcmjava.javaeditor.java2pcm.legacy
 Bundle-Vendor: vitruv.tools
