+++ tests/tools.vitruv.applications.pcmjava.javaeditor.tests/src/tools/vitruv/applications/pcmjava/javaeditor/java2pcm/ClassMappingTransformationTest.java
@@ -1,164 +1,118 @@
 package tools.vitruv.applications.pcmjava.javaeditor.java2pcm;
 
-import static org.junit.jupiter.api.Assertions.fail;
+import static edu.kit.ipd.sdq.commons.util.java.lang.IterableUtil.claimOne;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+import static tools.vitruv.applications.pcmjava.javaeditor.util.JavaQueryUtil.claimPackage;
+import static tools.vitruv.applications.pcmjava.javaeditor.util.PcmQueryUtil.claimComponent;
+import static tools.vitruv.applications.pcmjava.javaeditor.util.PcmQueryUtil.claimSingleRepository;
+import static tools.vitruv.applications.pcmjava.javaeditor.util.PcmQueryUtil.claimSingleSystem;
 
-import org.eclipse.emf.ecore.EObject;
+import java.util.stream.Collectors;
+
+import org.emftext.language.java.containers.Package;
 import org.junit.jupiter.api.Test;
 import org.palladiosimulator.pcm.repository.BasicComponent;
-import org.palladiosimulator.pcm.repository.CollectionDataType;
 import org.palladiosimulator.pcm.repository.CompositeComponent;
-import org.palladiosimulator.pcm.repository.CompositeDataType;
 import org.palladiosimulator.pcm.repository.Repository;
 import org.palladiosimulator.pcm.system.System;
 
 import tools.vitruv.applications.pcmjava.java2pcm.Java2PcmUserSelection;
 import tools.vitruv.applications.pcmjava.pcm2java.Pcm2JavaTestUtils;
 
-public class ClassMappingTransformationTest extends Java2PcmPackageMappingTransformationTest {
-
+public class ClassMappingTransformationTest extends Java2PcmTransformationTest {
 	/**
 	 * Class that in mapped package and same name as component + impl--> should be
 	 * the new implementing class for the component
-	 *
-	 * @throws Exception
 	 */
 	@Test
-	public void testAddComponentClassInPackageWithCorrespondingComponent() throws Throwable {
-		final Repository repo = super.addRepoContractsAndDatatypesPackage();
-		final BasicComponent bc = super.addSecondPackageCorrespondsToBasicComponent();
-
-		this.getUserInteraction().addNextSingleSelection(Java2PcmUserSelection.SELECT_BASIC_COMPONENT.getSelection());
-		final BasicComponent bcForClass = super.addClassInSecondPackage(BasicComponent.class);
-
-		super.assertRepositoryAndPCMName(repo, bcForClass, bc.getEntityName());
+	void testAddClassInSecondPackageAsBasicComponent() throws Exception {
+		createRepositoryPackage();
+		
+		getUserInteraction().addNextSingleSelection(Java2PcmUserSelection.SELECT_BASIC_COMPONENT.getSelection());
+		changeJavaView(view -> {
+			createPackageWithPackageInfo(view, Pcm2JavaTestUtils.REPOSITORY_NAME, Pcm2JavaTestUtils.BASIC_COMPONENT_NAME);
+		});
+		
+		getUserInteraction().addNextSingleSelection(Java2PcmUserSelection.SELECT_BASIC_COMPONENT.getSelection());
+		changeJavaEditorView(view -> {
+			Package innerPackage = claimPackage(view, Pcm2JavaTestUtils.BASIC_COMPONENT_NAME);
+			view.getManipulationUtil().createClass(Pcm2JavaTestUtils.BASIC_COMPONENT_NAME + Pcm2JavaTestUtils.IMPLEMENTING_CLASS_SUFFIX, innerPackage, null);
+		});
+		
+		validatePcmView(view -> {
+			Repository repository = claimSingleRepository(view);
+			claimComponent(repository, Pcm2JavaTestUtils.BASIC_COMPONENT_NAME, BasicComponent.class);
+		});
 	}
-
+	
 	@Test
-	public void testAddCompositeComponentClassInPackageWithCorrespondingCompositeComponent() throws Throwable {
-		final Repository repo = super.addRepoContractsAndDatatypesPackage();
-		final CompositeComponent cc = super.addSecondPackageCorrespondsToCompositeComponent();
-
-		this.getUserInteraction()
-				.addNextSingleSelection(Java2PcmUserSelection.SELECT_COMPOSITE_COMPONENT.getSelection());
-		final CompositeComponent ccForClass = this.addClassInSecondPackage(CompositeComponent.class);
-
-		super.assertRepositoryAndPCMName(repo, ccForClass, cc.getEntityName());
+	void testAddClassInSecondPackageAsCompositeComponent() throws Exception {
+		createRepositoryPackage();
+		
+		getUserInteraction().addNextSingleSelection(Java2PcmUserSelection.SELECT_COMPOSITE_COMPONENT.getSelection());
+		changeJavaView(view -> {
+			createPackageWithPackageInfo(view, Pcm2JavaTestUtils.REPOSITORY_NAME, Pcm2JavaTestUtils.COMPOSITE_COMPONENT_NAME);
+		});
+		
+		getUserInteraction().addNextSingleSelection(Java2PcmUserSelection.SELECT_COMPOSITE_COMPONENT.getSelection());
+		changeJavaEditorView(view -> {
+			Package innerPackage = claimPackage(view, Pcm2JavaTestUtils.COMPOSITE_COMPONENT_NAME);
+			view.getManipulationUtil().createClass(Pcm2JavaTestUtils.COMPOSITE_COMPONENT_NAME + Pcm2JavaTestUtils.IMPLEMENTING_CLASS_SUFFIX, innerPackage, null);
+		});
+		
+		validatePcmView(view -> {
+			Repository repository = claimSingleRepository(view);
+			claimComponent(repository, Pcm2JavaTestUtils.COMPOSITE_COMPONENT_NAME, CompositeComponent.class);
+		});
 	}
-
+	
 	@Test
-	public void testAddSystemClassInPackageWithCorrespondingSystem() throws Throwable {
-		super.addRepoContractsAndDatatypesPackage();
-		final System pcmSystem = super.addSecondPackageCorrespondsToSystem();
-
-		this.getUserInteraction().addNextSingleSelection(Java2PcmUserSelection.SELECT_SYSTEM.getSelection());
-		final System systemForClass = super.addClassInSecondPackage(System.class);
-
-		super.assertPCMNamedElement(systemForClass, pcmSystem.getEntityName());
+	void testAddClassInSecondPackageAsSystem() throws Exception {
+		createRepositoryPackage();
+		
+		getUserInteraction().addNextSingleSelection(Java2PcmUserSelection.SELECT_SYSTEM.getSelection());
+		changeJavaView(view -> {
+			createPackageWithPackageInfo(view, Pcm2JavaTestUtils.REPOSITORY_NAME, Pcm2JavaTestUtils.SYSTEM_NAME);
+		});
+		
+		getUserInteraction().addNextSingleSelection(Java2PcmUserSelection.SELECT_SYSTEM.getSelection());
+		changeJavaEditorView(view -> {
+			Package innerPackage = claimPackage(view, Pcm2JavaTestUtils.SYSTEM_NAME);
+			view.getManipulationUtil().createClass(Pcm2JavaTestUtils.SYSTEM_NAME + Pcm2JavaTestUtils.IMPLEMENTING_CLASS_SUFFIX, innerPackage, null);
+		});
+		
+		validatePcmView(view -> {
+			System system = claimSingleSystem(view);
+			assertEquals(Pcm2JavaTestUtils.SYSTEM_NAME, system.getEntityName(), "system has wrong name");
+		});
 	}
-
+	
 	/**
 	 * Test ii) class in non corresponding package --> should not be mapped to a
 	 * Basic Component
-	 *
-	 * @throws Throwable
 	 */
 	@Test
-	public void testAddClassInPackageWithoutCorrespondingComponent() throws Throwable {
-		super.addRepoContractsAndDatatypesPackage();
-		super.addSecondPackageCorrespondsWithoutCorrespondences();
-
-		this.getUserInteraction()
-				.addNextSingleSelection(Java2PcmUserSelection.SELECT_NOTHING_DECIDE_LATER.getSelection());
-		try {
-			final EObject eObject = super.addClassInPackage(this.secondPackage, EObject.class);
-			fail("The class should not have any correspondences, but it has a correspondence to eObject: " + eObject);
-		} catch (final RuntimeException re) {
-			// expected exception
-		}
+	void addClassInSecondPackageAsNone() throws Exception {
+		createRepositoryPackage();
+		
+		getUserInteraction().addNextSingleSelection(Java2PcmUserSelection.SELECT_NOTHING_DECIDE_LATER.getSelection());
+		changeJavaView(view -> {
+			createPackageWithPackageInfo(view, Pcm2JavaTestUtils.REPOSITORY_NAME, Pcm2JavaTestUtils.BASIC_COMPONENT_NAME);
+		});
+		
+		getUserInteraction().addNextSingleSelection(Java2PcmUserSelection.SELECT_NOTHING_DECIDE_LATER.getSelection());
+		changeJavaEditorView(view -> {
+			Package innerPackage = claimPackage(view, Pcm2JavaTestUtils.BASIC_COMPONENT_NAME);
+			view.getManipulationUtil().createClass(Pcm2JavaTestUtils.BASIC_COMPONENT_NAME + Pcm2JavaTestUtils.IMPLEMENTING_CLASS_SUFFIX, innerPackage, null);
+		});
+		
+		validatePcmView(view -> {
+			Repository repository = claimSingleRepository(view);
+			assertTrue(repository.getComponents__Repository().isEmpty(), "there must not exist any contained Component");
+			assertTrue(view.getRootObjects(System.class).isEmpty(), "there must not exist any System");
+		});
 	}
-
-	@Test
-	public void testAddBasicComponentClassInPackageWithoutCorrespondence() throws Throwable {
-		final Repository repo = this.addRepoContractsAndDatatypesPackage();
-		super.addSecondPackageCorrespondsWithoutCorrespondences();
-
-		this.getUserInteraction().addNextSingleSelection(Java2PcmUserSelection.SELECT_BASIC_COMPONENT.getSelection());
-		final BasicComponent newBc = super.addClassInSecondPackage(BasicComponent.class);
-
-		super.assertRepositoryAndPCMName(repo, newBc, newBc.getEntityName());
-	}
-
-	@Test
-	public void testAddCompositeComponentClassInPackageWithoutCorrespondence() throws Throwable {
-		final Repository repo = this.addRepoContractsAndDatatypesPackage();
-		super.addSecondPackageCorrespondsWithoutCorrespondences();
-
-		this.getUserInteraction()
-				.addNextSingleSelection(Java2PcmUserSelection.SELECT_COMPOSITE_COMPONENT.getSelection());
-		final CompositeComponent cc = super.addClassInSecondPackage(CompositeComponent.class);
-
-		super.assertRepositoryAndPCMName(repo, cc, cc.getEntityName());
-	}
-
-	@Test
-	public void testAddSystemClassInPackageWithoutCorrespondence() throws Throwable {
-		this.addRepoContractsAndDatatypesPackage();
-		super.addSecondPackageCorrespondsWithoutCorrespondences();
-
-		this.getUserInteraction().addNextSingleSelection(Java2PcmUserSelection.SELECT_SYSTEM.getSelection());
-		final System pcmSystem = super.addClassInSecondPackage(System.class);
-
-		this.assertPCMNamedElement(pcmSystem, Pcm2JavaTestUtils.IMPLEMENTING_CLASS_NAME);
-	}
-
-	@Test
-	public void testAddCompositeDatatypeClassInDatatypePackage() throws Throwable {
-		final Repository repo = this.addRepoContractsAndDatatypesPackage();
-
-		final CompositeDataType cdt = this.addClassThatCorrespondsToCompositeDatatype();
-
-		this.assertRepositoryAndPCMNameForDatatype(repo, cdt, Pcm2JavaTestUtils.IMPLEMENTING_CLASS_NAME);
-	}
-
-	@Test
-	public void testAddCollectionDatatypeClassInDatatypePackage() throws Throwable {
-		final Repository repo = this.addRepoContractsAndDatatypesPackage();
-
-		this.getUserInteraction()
-				.addNextSingleSelection(Java2PcmUserSelection.SELECT_COMPOSITE_COMPONENT.getSelection());
-		final CollectionDataType collection = super.addClassInPackage(this.getDatatypesPackage(),
-				CollectionDataType.class);
-
-		this.assertRepositoryAndPCMNameForDatatype(repo, collection, Pcm2JavaTestUtils.IMPLEMENTING_CLASS_NAME);
-	}
-
-	@Test
-	public void testAddClassInDatatypePackage() throws Throwable {
-		this.addRepoContractsAndDatatypesPackage();
-		try {
-			this.getUserInteraction()
-					.addNextSingleSelection(Java2PcmUserSelection.SELECT_NOTHING_DECIDE_LATER.getSelection());
-			final EObject eObject = super.addClassInPackage(this.getDatatypesPackage(), EObject.class);
-			fail("The class should not have any datatype correspondences, but it has a correspondence to eObject: "
-					+ eObject);
-		} catch (final RuntimeException re) {
-			// expected Exception
-		}
-	}
-
-	@Test
-	public void testRenameBasicComponentClass() throws Throwable {
-		final Repository repo = this.addRepoContractsAndDatatypesPackage();
-		this.addSecondPackageCorrespondsWithoutCorrespondences();
-		this.getUserInteraction().addNextSingleSelection(Java2PcmUserSelection.SELECT_BASIC_COMPONENT.getSelection());
-		final BasicComponent basicComponent = this.addClassInSecondPackage(BasicComponent.class);
-
-		final BasicComponent newBasicComponent = super.renameClassifierWithName(basicComponent.getEntityName() + "Impl",
-				Pcm2JavaTestUtils.IMPLEMENTING_CLASS_NAME + Pcm2JavaTestUtils.RENAME, BasicComponent.class);
-
-		this.assertRepositoryAndPCMName(repo, newBasicComponent,
-				Pcm2JavaTestUtils.IMPLEMENTING_CLASS_NAME + Pcm2JavaTestUtils.RENAME);
-		this.assertFilesOnlyForEObjects(newBasicComponent);
-	}
-
+	
+	//TODO: add more
 }
