--- tests/tools.vitruv.applications.pcmjava.javaeditor.tests/src/tools/vitruv/applications/pcmjava/javaeditor/java2pcm/SynchronizationAwaitCallback.java
@@ -1,5 +1,0 @@
-package tools.vitruv.applications.pcmjava.javaeditor.java2pcm;
-
-public interface SynchronizationAwaitCallback {
-	void waitForSynchronization(int numberOfExpectedSynchronizationCalls);
-}
